+++ .gitignore
@@ -8,4 +8,4 @@
 *.d
 tmp.*
 temp.*
-__pycache__
+__pycache*
