+++ Readme.md
@@ -1,5 +1,9 @@
 # ANoobs Brres Material Tool (ABMatT)
 This command line tool is for editing materials in _Brres_ files in _Mario Kart Wii_. Python is required to run it. The tool can do quick edits from the command line, or read in a command file for processing multiple setting adjustments. This is particularly useful for editing a large amount of materials or recreating a brres multiple times. Python regular expression matching is supported.
+
+## Known Limitations
+ABMatT currently only supports overwriting data, not additions or deletions.
+
 ## Transparency
 The tool is also smart about adjusting transparency. When setting materials to transparent it also updates the comparison and reference settings, and the draw list to be xlu (fixing Harry Potter effect). These settings can also be adjusted individually, including draw priority.
