+++ brres/testcmd.txt
@@ -1,0 +1,1 @@
+info shader
