+++ brres/testcmd2.txt
@@ -1,0 +1,1 @@
+set shader:2 for *
