+++ brres/todo.md
@@ -1,15 +1,8 @@
 # Todo
-* Read/Write Brres file
-* Command line input
-* Set materials transparent as a group
-* Set objects transparent as a group
-* Set Texrefs mipmap/scale settings as a group
-* Sort shaders
+* Testing of application
+* Fixing CTools Brres?
+* changing texture references
+* Adding and deleting from file (shaders and layers)
+* Command line interactive mode
 * Set material/shader/animation ef_dushboard and other common materials
 * Shadow material/shader
-
-
-## Dependencies
-* Input: Parsing format decisions for program control
-* Data structures both for manipulation and for file operations
-* Output: How to display data
